Remove unsafe pointer in favor of macros :unamused:
The class does not have enough uses for it to warrant its own
existence. It is also not really possible to use in functions marked
as generic since gcc will not inline any of its methods. This is a big
concern since the whole point of this class was for it to be useful.