Extend memset for multibyte :chains:
Filling a string with bytes is of course important, but so is also
filling a string with code points. Since utf8 (and friends) however
are variable length encodings, we need the ability to set multiple
bytes at a time. This change extends memset to accept generic
characters as the filling.