Link ground truth conversion library to common :chains:
There is little point of duplicating the test/benchmark data to all
unicode encoding forms, better storing it in one small form (utf8) and
convert it to the others when needed. This, however, forces us to rely
on some algorithm to be the ground truth and since we do not trust our
own conversion algorithms we have to use some other library.