Remove generic target :red_circle:
The generic target causes a whole bunch of problems with inline of
helper/intrinsics functions. The macro started out as a hack in order
to solve a problem with gcc generating simd instructions for
algorithms that specifically were written without such support, and
badly at that since they crashed at runtime. This is probably no
longer any problem because any such algorithms ought to be placed in
code files and compiled without knowledge of any simd
instructions.