Avoid build collision with parent project :family:
Having targets named `lib_` is too generic. We can pretend that the
prefix `ful_` is reserved by this library.

The working directory for tests is also changed rather than changing
the hardcoded path to the data directory in the source (see
`cmn/data.hpp`.)

Similarly, the build property `DEP_DIRECTORY` is only set if not set
already.