Optimize strlen :stopwatch:
Turns out that glibc's implementation of strlen for 8-bit characters
is really good. Like, seriously impressive. At first, instead of
rewriting their implementation, I decided to piggyback on them and use
it if available but since this library also needed implementations for
16- and 32-bit characters it made more sense to not do that.

The only really difference in implementation between this and the one
in glibc is that the entry point for this version of strlen is inlined
for short strings (depending on the vector size available) that do not
cross a page boundary. It is also slightly slower on medium sized
strings, maybe, it is difficult to tell for sure.